support for nose with gratings

+++ python/src/kernelphysiology/dl/pytorch/utils/cv2_functional.py
@@ -45,7 +45,8 @@ def to_tensor(pic):
 
     if _is_numpy_image(pic):
         if len(pic.shape) == 2:
-            pic = cv2.cvtColor(pic, cv2.COLOR_GRAY2RGB)
+            # FIXME: why not accepting a channel iamges?
+            pic = cv2.cvtColor(pic.astype('float32'), cv2.COLOR_GRAY2RGB)
         img = torch.from_numpy(pic.transpose((2, 0, 1))).type(torch.FloatTensor)
         # backward compatibility
         if isinstance(img, torch.ByteTensor) or img.max() > 1:
